requirement.txt
tensorflow==2.16.1
numpy==1.25.0
Pillow
scipy==1.11.1
streamlit
matplotlib==3.7.1


